Work In Progress with a teardown redo
